rename view name
