+++ src/renderer/App.vue
@@ -6,7 +6,7 @@
     <ResearchDialog v-if="dialogVisibilities.research" />
     <AnalysisDialog v-if="dialogVisibilities.analysis" />
     <USIEngineManagementDialog v-if="dialogVisibilities.usiEngineSetting" />
-    <ExportBoardImageDialog v-if="dialogVisibilities.exportPositionImage" />
+    <ExportPositionImageDialog v-if="dialogVisibilities.exportPositionImage" />
     <AppSettingDialog v-if="dialogVisibilities.appSetting" />
     <PasteDialog v-if="dialogVisibilities.paste" />
     <BussyMessage v-if="dialogVisibilities.bussy" />
@@ -24,7 +24,7 @@ import GameDialog from "@/renderer/view/dialog/GameDialog.vue";
 import CSAGameDialog from "@/renderer/view/dialog/CSAGameDialog.vue";
 import ResearchDialog from "@/renderer/view/dialog/ResearchDialog.vue";
 import USIEngineManagementDialog from "@/renderer/view/dialog/USIEngineManagementDialog.vue";
-import ExportBoardImageDialog from "@/renderer/view/dialog/ExportBoardImageDialog.vue";
+import ExportPositionImageDialog from "@/renderer/view/dialog/ExportPositionImageDialog.vue";
 import AppSettingDialog from "@/renderer/view/dialog/AppSettingDialog.vue";
 import PasteDialog from "@/renderer/view/dialog/PasteDialog.vue";
 import BussyMessage from "@/renderer/view/dialog/BussyMessage.vue";
@@ -46,7 +46,7 @@ export default defineComponent({
     CSAGameDialog,
     ResearchDialog,
     USIEngineManagementDialog,
-    ExportBoardImageDialog,
+    ExportPositionImageDialog,
     AppSettingDialog,
     PasteDialog,
     BussyMessage,
