+++ src/renderer/view/dialog/ExportPositionImageDialog.vue
